[feature/asset-domain]fix: 오류수정 1.잘못 표기한 type 변경 Asset (Integer -> Long) / contentId (Integer -> Long) 2.UUID random 으로 변경

+++ lms/src/main/java/com/lms/domain/asset/creation/CreateAsset.java
@@ -4,5 +4,5 @@ public record CreateAsset(
     String mimeType,
     String path,
     String originalFilename,
-    Integer contentId
+    Long contentId
 ) {}
